requirements.txt
suanpan[docker]==0.9.87
tensorflow
imageio
pillow